+++ app/src/main/java/com/example/trabalhofinal/fragments/FragmentLogado.java
@@ -1,4 +1,4 @@
-package com.example.trabalhofinal.Fragments;
+package com.example.trabalhofinal.fragments;
 
 import android.os.Bundle;
