+++ app/src/main/res/layout/fragment_lista_list.xml
@@ -9,5 +9,5 @@
     android:layout_marginLeft="16dp"
     android:layout_marginRight="16dp"
     app:layoutManager="LinearLayoutManager"
-    tools:context=".Fragments.FragmentLista"
+    tools:context=".fragments.FragmentLista"
     tools:listitem="@layout/fragment_lista" />
