+++ app/src/main/res/layout/fragment_rastreio.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Fragments.FragmentRastreio">
+    tools:context=".fragments.FragmentRastreio">
 
     <com.google.android.material.textfield.TextInputEditText
         android:id="@+id/inputcodigo"
