+++ app/src/main/res/layout/fragmento_home.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Fragments.FragmentHome">
+    tools:context=".fragments.FragmentHome">
 
     <Button
         android:id="@+id/btnrastreio"
